extract helper to open account for editing
