+++ deployables/extension/src/panel/accounts/index.ts
@@ -1,3 +1,4 @@
+export { editAccount } from './editAccount'
 export { getAccount } from './getAccount'
 export { getAccounts } from './getAccounts'
 export { getActiveAccount } from './getActiveAccount'
